requirements.txt
google-genai>=1.0.0
pypdf>=4.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
tiktoken
agentlightning>=0.3.0
datasets>=2.14.0
slowapi>=0.1.9
python-magic>=0.4.27
redis>=5.0.0

pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0
